+++ Sources/ZcashLightClientKit/ClosureSynchronizer.swift
@@ -128,7 +128,9 @@ public protocol ClosureSynchronizer {
     func refreshUTXOs(address: TransparentAddress, from height: BlockHeight, completion: @escaping (Result<RefreshedUTXOs, Error>) -> Void)
 
     func getAccountBalance(accountIndex: Int, completion: @escaping (Result<AccountBalance?, Error>) -> Void)
-    
+
+    func refreshExchangeRateUSD()
+
     /*
      It can be missleading that these two methods are returning Publisher even this protocol is closure based. Reason is that Synchronizer doesn't
      provide different implementations for these two methods. So Combine it is even here.