+++ CHANGELOG.md
@@ -6,6 +6,29 @@ and this library adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Unreleased
 
+## Added
+- `Synchronizer.exchangeRateUSDStream: AnyPublisher<FiatCurrencyResult?, Never>`,
+  which returns the currently-cached USD/ZEC exchange rate, or `nil` if it has not yet been
+  fetched.
+- `Synchronizer.refreshExchangeRateUSD()`, which refreshes the rate returned by
+  `Synchronizer.exchangeRateUSDStream`. Prices are queried over Tor (to hide the wallet's
+  IP address).
+
+## Changed
+
+### [#1475] Adopt transaction data requests
+- The transaction history is now processed using `transaction data requests`, which are fetched every 1,000 blocks during longer syncs or with each sync loop when a new block is mined.
+
+## Checkpoints
+
+Mainnet
+
+````
+Sources/ZcashLightClientKit/Resources/checkpoints/mainnet/2562500.json
+...
+Sources/ZcashLightClientKit/Resources/checkpoints/mainnet/2610000.json
+````
+
 # 2.1.12 - 2024-07-04
 
 ## Fixed